Use a different format for ContestExporter/Importer
The new format isn't "tree-like" anymore but, instead, it's a dictionary
whose values are all exported objects and whose keys are the associated
IDs. The IDs aren't the ones they had in the DB before exportation nor
the ones they will have after importation. They're IDs whose scope is
just the one of the exported data, and they're global IDs (unique among
all classes). This allows us to export complex relationships, and not
only trees. Moreover we are able to write a general method that does
this for all classes and all relationships, avoiding the hacks we used,
for example, when exporting submissions.

With this commit ContestExporter doesn't use .export_to_dict() methods
anymore and ContestImporter doesn't use .import_from_dict().